==> 01b-analysis.R <==
getwd()
setwd("/Users/eylulaygun/Desktop/Year\ 5/stat\ 447B/CovidSentimentAnalysis")


# read dataframe and perform basic transformations
df <- read.csv("updated_dataframe-apr06.csv")

#' Create new variable, popularity_score, which is the response variable for project
#' popularity score is a weighted average of number of retweets and favorites
#' 
#' Intuitively, growth in reach generated by retweets is not linear but exponential (just like spreading covid!),
#  and with slight increments in retweets we can still achieve large increases in the reach of a tweet, 
#' which is what makes a tweet go popular
#'
#'@param data entire dataframe
#'@param weights List[2] of numbers. weights to apply for retweets and favorites columns
#'@returns popularity score
#'
GetPopularityScore <-function(data, weighs){
  return (weighs[1]*data$retweets + weighs[2]*data$favorites)
}

#' Transforms given column into ordinal categorical 
#' @param data Column to transform, needs to be type int
#' @param breaks List of int, Breaks for transformation
#' @param labels List of Char, labels for each category
#' @returns a new column which is ordinal categorical
TransformIntoOrdinalCategorical <-function(data, breaks, labels){
  return (cut(data, breaks = breaks, labels = labels, ordered_results = TRUE, right = TRUE ))
}

# Data Wrangling
colnames(df)
df <- df[c("user_name","user_verified", "date", "text", "hashtags", "hashtags_count", "retweets", "favorites","compound_score", 
           "positive_score", "negative_score", "neutral_score")]

# creating response variable
weighs = c(0.75, 0.25)
df$popularity_score <- GetPopularityScore(df, weighs = weighs) 
df$popularity_score_ctg <- TransformIntoOrdinalCategorical(data = df$popularity_score, breaks =  c(-0.1, 1, 10, 1000), labels = c("LOW", "AVG", "HIGH"))

# creating explanatory variables
df$overall_sentiment <- TransformIntoOrdinalCategorical(data= df$compound_score, breaks = c(-1, -0.05, 0.05, 1), labels =  c("NEG", "NEU", "POS"))
df$date <- as.Date(df$date,tz=Sys.timezone(), format ="%Y-%m-%d") # remove H-M-Sec from date for simplicity
df$hashtags_count <- TransformIntoOrdinalCategorical(data = df$hashtags_count, breaks = c(-1, 1, 4, 11), labels = c(1, 2, 3))
df$tweet_age <- as.numeric(df$date - as.Date(as.character("2020-12-12", format = "%Y-%m-%d")))
# save new csv 
write.csv(df, "01b-updated_dataframe-apr06.csv")
